fixed implicit conversion warning in renderer2d

+++ Ares/src/Ares/Renderer/Renderer2D.cpp
@@ -191,7 +191,7 @@ namespace Ares
 		ARES_PROFILE_FUNCTION();
 
 
-		uint32_t dataSize = (uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase;
+		uint32_t dataSize = (uint32_t)((uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase);
 		s_Data.QuadVertexBuffer->SetData(s_Data.QuadVertexBufferBase, dataSize);
